requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
python-jose
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
pydantic-settings